==> 0 - extend_dataset.ipynb <==
from modules.data_loader import DataLoaderPM100

loader = DataLoaderPM100("datasets/pm100/job_table.parquet")

df = loader.load_data()
df.head()

df = loader.extract_additional_features()
df.head()

train_df, test_df = loader.split_train_test()

df.shape

train_df.shape

test_df.shape

loader.save_parquet("job_data")

import pandas as pd

df_full = pd.read_parquet("datasets/pm100/job_data_full.parquet", engine="pyarrow").copy()
df_full.head()

df_full.columns

df_train = pd.read_parquet("datasets/pm100/job_data_train.parquet", engine="pyarrow").copy()
df_train.head()

df_test = pd.read_parquet("datasets/pm100/job_data_test.parquet", engine="pyarrow").copy()
df_test.head()
