==> 4 - KNN.ipynb <==
## Preliminaries and Utils

import numpy as np
import pandas as pd
import warnings
import time

from sklearn.cluster import KMeans
from sklearn.preprocessing import  MinMaxScaler, StandardScaler

# from modules.prediction_models import OnlineDecisionTreeRegressor
# from modules.prediction_models import OnlineRidgePolynomialRegressor
from modules.prediction_models import OnlineKNNRegressor

from modules.utils import Metrics, PrintSummary, ShowPlots

warnings.filterwarnings('ignore', category=DeprecationWarning)
warnings.filterwarnings('ignore', category=FutureWarning)
warnings.filterwarnings("ignore", message="X does not have valid feature names")
warnings.filterwarnings("ignore", message="X has feature names, but PolynomialFeatures was fitted without feature names")
warnings.filterwarnings("ignore", message="X has feature names, but DecisionTreeRegressor was fitted without feature names")
pd.options.mode.chained_assignment = None

summary = PrintSummary()
plots = ShowPlots()
metrics = Metrics()

## Data Loading

data_train = 'datasets/pm100/job_data_train.parquet'
data_test = 'datasets/pm100/job_data_test.parquet'

df_train = pd.read_parquet(data_train, engine="pyarrow").copy()
df_test = pd.read_parquet(data_test, engine="pyarrow").copy()

df_train['Desired QoS'] = df_train['Desired QoS'].astype(int)
df_test['Desired QoS'] = df_test['Desired QoS'].astype(int)

len_test = len(df_test)
len_test

## Data Preparation

scaler = MinMaxScaler()
scaler.fit(df_train['Run Time'].values.reshape(-1, 1))

fs1 = ["User ID", "Requested Number of Nodes", "Requested Number of CPU", "Requested Number of GPU", "Total Requested Memory", "Desired QoS", "Requested Time"]

fs2 = ["User ID", "Requested Number of Nodes", "Requested Number of CPU", "Requested Number of GPU", "Total Requested Memory", "Desired QoS", "Requested Time",
       "Prev Run Time 1", "Prev Run Time 2", "Prev Run Time 3", "Avg Run Time 2", "Avg Run Time 3", "Avg Run Time All"]

fs3 = ["User ID", "Requested Number of Nodes", "Requested Number of CPU", "Requested Number of GPU", "Total Requested Memory", "Desired QoS", "Requested Time",
       "Prev Run Time 1", "Prev Run Time 2", "Prev Run Time 3", "Avg Run Time 2", "Avg Run Time 3", "Avg Run Time All",
       "Prev Requested Nodes 1", "Prev Requested Nodes 2", "Prev Requested Nodes 3", "Avg Requested Nodes 2", "Avg Requested Nodes 3", "Avg Requested Nodes All"]

target_name = "Run Time"

y_train = df_train[target_name].values
y_test  = df_test[target_name]

## SET 1

X_train = df_train[fs1].values
X_test  = df_test[fs1].values

val_size = int(len(df_train) * 0.1)

# Consecutive split
df_train_sub = df_train.iloc[:-val_size]
df_val       = df_train.iloc[-val_size:]

# Targets
y_train_sub = df_train_sub[target_name].astype(float).values
y_val       = df_val[target_name].astype(float).values

# Features
X_train_sub = df_train_sub[fs1].astype(float).values
X_val       = df_val[fs1].astype(float).values

from sklearn.metrics import mean_absolute_error
from tqdm import tqdm

coarse_k = list(range(5, 201, 10))
coarse_results = []

for k in tqdm(coarse_k, desc="Coarse k search"):
    reg = OnlineKNNRegressor(k=k)
    reg.partial_fit(X_train_sub, y_train_sub)
    y_val_pred = reg.predict(X_val)
    mae = mean_absolute_error(y_val, y_val_pred)
    coarse_results.append((k, mae))

best_k_coarse, _ = min(coarse_results, key=lambda x: x[1])

refine_range = range(
    max(3, best_k_coarse - 10),
    best_k_coarse + 11
)

refined_results = []

for k in tqdm(refine_range, desc="Refining k"):
    reg = OnlineKNNRegressor(k=k)
    reg.partial_fit(X_train_sub, y_train_sub)
    y_val_pred = reg.predict(X_val)
    mae = mean_absolute_error(y_val, y_val_pred)
    refined_results.append((k, mae))

k_opt, min_mae = min(refined_results, key=lambda x: x[1])

print(f"Optimal k = {k_opt}, Validation MAE = {min_mae}")

regressor = OnlineKNNRegressor(k=k_opt)
regressor.partial_fit(X_train, y_train)

start = time.time()
y_pred = regressor.predict(X_test)
end = time.time()

metrics.print(scaler, y_test, y_pred, start, end, len_test)

# Save the results in the file with all other predictions

df = pd.read_csv("predictions/pm100/predictions.csv")

df["pred_runtime_knn_fs1"] = y_pred

df.to_csv("predictions/pm100/predictions.csv", index=False)

## SET 2

X_train = df_train[fs2].values
X_test  = df_test[fs2].values

val_size = int(len(df_train) * 0.1)

# Consecutive split
df_train_sub = df_train.iloc[:-val_size]
df_val       = df_train.iloc[-val_size:]

# Targets
y_train_sub = df_train_sub[target_name].astype(float).values
y_val       = df_val[target_name].astype(float).values

# Features
X_train_sub = df_train_sub[fs2].astype(float).values
X_val       = df_val[fs2].astype(float).values

from sklearn.metrics import mean_absolute_error
from tqdm import tqdm

coarse_k = list(range(5, 201, 10))
coarse_results = []

for k in tqdm(coarse_k, desc="Coarse k search"):
    reg = OnlineKNNRegressor(k=k)
    reg.partial_fit(X_train_sub, y_train_sub)
    y_val_pred = reg.predict(X_val)
    mae = mean_absolute_error(y_val, y_val_pred)
    coarse_results.append((k, mae))

best_k_coarse, _ = min(coarse_results, key=lambda x: x[1])

refine_range = range(
    max(3, best_k_coarse - 10),
    best_k_coarse + 11
)

refined_results = []

for k in tqdm(refine_range, desc="Refining k"):
    reg = OnlineKNNRegressor(k=k)
    reg.partial_fit(X_train_sub, y_train_sub)
    y_val_pred = reg.predict(X_val)
    mae = mean_absolute_error(y_val, y_val_pred)
    refined_results.append((k, mae))

k_opt, min_mae = min(refined_results, key=lambda x: x[1])

print(f"Optimal k = {k_opt}, Validation MAE = {min_mae}")

regressor = OnlineKNNRegressor(k=k_opt)
regressor.partial_fit(X_train, y_train)

start = time.time()
y_pred = regressor.predict(X_test)
end = time.time()

metrics.print(scaler, y_test, y_pred, start, end, len_test)

# Save the results in the file with all other predictions

df = pd.read_csv("predictions/pm100/predictions.csv")

df["pred_runtime_knn_fs2"] = y_pred

df.to_csv("predictions/pm100/predictions.csv", index=False)
