==> 7 - fix_predictions.ipynb <==
import numpy as np
import pandas as pd

# Load CSV
input_csv = "predictions/pm100/predictions.csv"

df = pd.read_csv(input_csv)

df.dtypes

df[df.pred_runtime_heuristic > 86400]

df[df.pred_runtime_heuristic > df.pred_runtime_user]

df["prev_pred"] = df.groupby("user_id")["pred_runtime_heuristic"].shift(1)
df["prev_gt"] = df.groupby("user_id")["gt_runtime"].shift(1)

df["pred_ratio"] = df["pred_runtime_heuristic"] / df["prev_pred"]
df["gt_ratio"] = df["prev_gt"] / df["prev_pred"]

PRED_JUMP = 100     # prediction exploded
GT_JUMP   = 250      # real runtime stayed reasonable
MIN_PREV  = 10      # avoid division noise

df["suspicious"] = (
    ((df["pred_ratio"] > PRED_JUMP) & 
    (df["gt_ratio"] <= GT_JUMP))
)

cols = [
    "user_id", "time_submit",
    "prev_pred", "pred_runtime_heuristic", "pred_ratio",
    "prev_gt", "gt_runtime", "gt_ratio"
]

df.loc[df["suspicious"], cols].head(20)

df.loc[df["suspicious"], "pred_runtime_heuristic"] /= 1000

df

df.drop(columns=["prev_pred"], inplace=True)
df.drop(columns=["prev_gt"], inplace=True)
df.drop(columns=["pred_ratio"], inplace=True)
df.drop(columns=["gt_ratio"], inplace=True)
df.drop(columns=["suspicious"], inplace=True)

df

# Convert to integer
df["pred_runtime_heuristic"] = (
    df["pred_runtime_heuristic"]
    .round()
    .astype(int)
)

df["pred_runtime_dt_fs1"] = (
    df["pred_runtime_dt_fs1"]
    .round()
    .astype(int)
)

df["pred_runtime_dt_fs2"] = (
    df["pred_runtime_dt_fs2"]
    .round()
    .astype(int)
)

df["pred_runtime_rnp_fs1"] = (
    df["pred_runtime_rnp_fs1"]
    .round()
    .astype(int)
)

df["pred_runtime_rnp_fs2"] = (
    df["pred_runtime_rnp_fs2"]
    .round()
    .astype(int)
)

df["pred_runtime_knn_fs1"] = (
    df["pred_runtime_knn_fs1"]
    .round()
    .astype(int)
)

df["pred_runtime_knn_fs2"] = (
    df["pred_runtime_knn_fs2"]
    .round()
    .astype(int)
)

df["pred_runtime_ora"] = (
    df["pred_runtime_ora"]
    .round()
    .astype(int)
)

# Cap predictions to user prediction
df["pred_runtime_heuristic"] = np.minimum(
    df["pred_runtime_heuristic"],
    df["pred_runtime_user"]
)

df["pred_runtime_dt_fs1"] = np.minimum(
    df["pred_runtime_dt_fs1"],
    df["pred_runtime_user"]
)

df["pred_runtime_dt_fs2"] = np.minimum(
    df["pred_runtime_dt_fs2"],
    df["pred_runtime_user"]
)

df["pred_runtime_rnp_fs1"] = np.minimum(
    df["pred_runtime_rnp_fs1"],
    df["pred_runtime_user"]
)

df["pred_runtime_rnp_fs2"] = np.minimum(
    df["pred_runtime_rnp_fs2"],
    df["pred_runtime_user"]
)

df["pred_runtime_knn_fs1"] = np.minimum(
    df["pred_runtime_knn_fs1"],
    df["pred_runtime_user"]
)

df["pred_runtime_knn_fs2"] = np.minimum(
    df["pred_runtime_knn_fs2"],
    df["pred_runtime_user"]
)

df["pred_runtime_ora"] = np.minimum(
    df["pred_runtime_ora"],
    df["pred_runtime_user"]
)

df

# Save corrected CSV
df.to_csv(input_csv, index=False)

print(f"Corrected file saved to: {input_csv}")
