==> 8 - compute_metrics.ipynb <==
## Preliminaries

# pip install matplotlib pandas scikit-learn seaborn

import matplotlib.pyplot as plt
import numpy as np
import pandas as pd
import seaborn as sns

from sklearn.metrics import mean_absolute_error

from modules.utils import PrintSummary, ShowPlots

## Data Loading

predictions = pd.read_csv("predictions/pm100/predictions.csv")
predictions

w_900_1 = pd.read_csv('workloads/pm100/pm100_900_1.swf', sep="\t", header=None)
w_900_2 = pd.read_csv('workloads/pm100/pm100_900_2.swf', sep="\t", header=None)
w_900_3 = pd.read_csv('workloads/pm100/pm100_900_3.swf', sep="\t", header=None)
w_900_4 = pd.read_csv('workloads/pm100/pm100_900_4.swf', sep="\t", header=None)
w_900_5 = pd.read_csv('workloads/pm100/pm100_900_5.swf', sep="\t", header=None)

w_9k_1 = pd.read_csv('workloads/pm100/pm100_9k_1.swf', sep="\t", header=None)
w_9k_2 = pd.read_csv('workloads/pm100/pm100_9k_2.swf', sep="\t", header=None)
w_9k_3 = pd.read_csv('workloads/pm100/pm100_9k_3.swf', sep="\t", header=None)
w_9k_4 = pd.read_csv('workloads/pm100/pm100_9k_4.swf', sep="\t", header=None)
w_9k_5 = pd.read_csv('workloads/pm100/pm100_9k_5.swf', sep="\t", header=None)

## Data Filtering

pred_w_900_1 = predictions[predictions["job_id"].isin(w_900_1[0].values)]
pred_w_900_2 = predictions[predictions["job_id"].isin(w_900_2[0].values)]
pred_w_900_3 = predictions[predictions["job_id"].isin(w_900_3[0].values)]
pred_w_900_4 = predictions[predictions["job_id"].isin(w_900_4[0].values)]
pred_w_900_5 = predictions[predictions["job_id"].isin(w_900_5[0].values)]

pred_w_9k_1 = predictions[predictions["job_id"].isin(w_9k_1[0].values)]
pred_w_9k_2 = predictions[predictions["job_id"].isin(w_9k_2[0].values)]
pred_w_9k_3 = predictions[predictions["job_id"].isin(w_9k_3[0].values)]
pred_w_9k_4 = predictions[predictions["job_id"].isin(w_9k_4[0].values)]
pred_w_9k_5 = predictions[predictions["job_id"].isin(w_9k_5[0].values)]

## Metrics Helpers

def mean_average_percentage_error(y_true, y_pred):
    y_true = np.array(y_true, dtype=float)
    y_pred = np.array(y_pred, dtype=float)

    # Avoid division by zero
    mask = (y_true > 0) & (y_pred > 0)
    y_true = y_true[mask]
    y_pred = y_pred[mask]
    
    mape = np.mean(np.abs((y_true - y_pred) / y_true)) * 100

    return mape

def estimation_accuracy(y_true, y_pred):
    y_true = np.array(y_true, dtype=float)
    y_pred = np.array(y_pred, dtype=float)

    # Avoid division by zero
    mask = (y_true > 0) & (y_pred > 0)
    y_true = y_true[mask]
    y_pred = y_pred[mask]

    ea = np.where(y_pred <= y_true,
                  y_pred / y_true,
                  y_true / y_pred)

    return float(np.mean(ea))

def convert_to_hms(time_value, unit="seconds"):
    if unit == "minutes":
        time_value *= 60  # Convert minutes to seconds
    hours, remainder = divmod(time_value, 3600)
    minutes, seconds = divmod(remainder, 60)
    return f"{int(hours):02}:{int(minutes):02}:{int(seconds):02}"

# gt_runtime pred_runtime_user pred_runtime_heuristic pred_runtime_dt_fs2 pred_runtime_rnp_fs2 pred_runtime_knn_fs2 pred_runtime_ora

## Complete Dataset

### Metric Computation

mae_usr = mean_absolute_error(predictions.gt_runtime, predictions.pred_runtime_user)
mae_heu = mean_absolute_error(predictions.gt_runtime, predictions.pred_runtime_heuristic)
mae_dt  = mean_absolute_error(predictions.gt_runtime, predictions.pred_runtime_dt_fs2)
mae_rnp = mean_absolute_error(predictions.gt_runtime, predictions.pred_runtime_rnp_fs2)
mae_knn = mean_absolute_error(predictions.gt_runtime, predictions.pred_runtime_knn_fs2)
mae_llm = mean_absolute_error(predictions.gt_runtime, predictions.pred_runtime_ora)

ea_usr = estimation_accuracy(predictions.gt_runtime, predictions.pred_runtime_user)
ea_heu = estimation_accuracy(predictions.gt_runtime, predictions.pred_runtime_heuristic)
ea_dt  = estimation_accuracy(predictions.gt_runtime, predictions.pred_runtime_dt_fs2)
ea_rnp = estimation_accuracy(predictions.gt_runtime, predictions.pred_runtime_rnp_fs2)
ea_knn = estimation_accuracy(predictions.gt_runtime, predictions.pred_runtime_knn_fs2)
ea_llm = estimation_accuracy(predictions.gt_runtime, predictions.pred_runtime_ora)

mape_usr = mean_average_percentage_error(predictions.gt_runtime, predictions.pred_runtime_user)
mape_heu = mean_average_percentage_error(predictions.gt_runtime, predictions.pred_runtime_heuristic)
mape_dt  = mean_average_percentage_error(predictions.gt_runtime, predictions.pred_runtime_dt_fs2)
mape_rnp = mean_average_percentage_error(predictions.gt_runtime, predictions.pred_runtime_rnp_fs2)
mape_knn = mean_average_percentage_error(predictions.gt_runtime, predictions.pred_runtime_knn_fs2)
mape_llm = mean_average_percentage_error(predictions.gt_runtime, predictions.pred_runtime_ora)

### Metrics

res = pd.DataFrame()
res['Workload'] = ('C', 'C', 'C', 'C', 'C', 'C')
res['Method'] = ('User', 'Heuristic', 'DT', 'RNP', 'KNN', 'ORA')
res['MAE'] = (mae_usr, mae_heu, mae_dt, mae_rnp, mae_knn, mae_llm)
res['Average Error'] = (convert_to_hms(mae_usr), convert_to_hms(mae_heu), convert_to_hms(mae_dt), convert_to_hms(mae_rnp), convert_to_hms(mae_knn), convert_to_hms(mae_llm))
res['EA'] = (ea_usr, ea_heu, ea_dt, ea_rnp, ea_knn, ea_llm)
res['MAPE'] = (mape_usr/1000, mape_heu/1000, mape_dt/1000, mape_rnp/1000, mape_knn/1000, mape_llm/1000)

res

for col in predictions.columns:
    max_col = max(predictions[col])
    print(f"{col} = {max_col}")

## Scenario (a): 900x98

### Metric Computation

mae_usr_sa1 = mean_absolute_error(pred_w_900_1.gt_runtime, pred_w_900_1.pred_runtime_user)
mae_heu_sa1 = mean_absolute_error(pred_w_900_1.gt_runtime, pred_w_900_1.pred_runtime_heuristic)
mae_dt_sa1  = mean_absolute_error(pred_w_900_1.gt_runtime, pred_w_900_1.pred_runtime_dt_fs2)
mae_rnp_sa1 = mean_absolute_error(pred_w_900_1.gt_runtime, pred_w_900_1.pred_runtime_rnp_fs2)
mae_knn_sa1 = mean_absolute_error(pred_w_900_1.gt_runtime, pred_w_900_1.pred_runtime_knn_fs2)
mae_llm_sa1 = mean_absolute_error(pred_w_900_1.gt_runtime, pred_w_900_1.pred_runtime_ora)

mae_usr_sa2 = mean_absolute_error(pred_w_900_2.gt_runtime, pred_w_900_2.pred_runtime_user)
mae_heu_sa2 = mean_absolute_error(pred_w_900_2.gt_runtime, pred_w_900_2.pred_runtime_heuristic)
mae_dt_sa2  = mean_absolute_error(pred_w_900_2.gt_runtime, pred_w_900_2.pred_runtime_dt_fs2)
mae_rnp_sa2 = mean_absolute_error(pred_w_900_2.gt_runtime, pred_w_900_2.pred_runtime_rnp_fs2)
mae_knn_sa2 = mean_absolute_error(pred_w_900_2.gt_runtime, pred_w_900_2.pred_runtime_knn_fs2)
mae_llm_sa2 = mean_absolute_error(pred_w_900_2.gt_runtime, pred_w_900_2.pred_runtime_ora)

mae_usr_sa3 = mean_absolute_error(pred_w_900_3.gt_runtime, pred_w_900_3.pred_runtime_user)
mae_heu_sa3 = mean_absolute_error(pred_w_900_3.gt_runtime, pred_w_900_3.pred_runtime_heuristic)
mae_dt_sa3  = mean_absolute_error(pred_w_900_3.gt_runtime, pred_w_900_3.pred_runtime_dt_fs2)
mae_rnp_sa3 = mean_absolute_error(pred_w_900_3.gt_runtime, pred_w_900_3.pred_runtime_rnp_fs2)
mae_knn_sa3 = mean_absolute_error(pred_w_900_3.gt_runtime, pred_w_900_3.pred_runtime_knn_fs2)
mae_llm_sa3 = mean_absolute_error(pred_w_900_3.gt_runtime, pred_w_900_3.pred_runtime_ora)

mae_usr_sa4 = mean_absolute_error(pred_w_900_4.gt_runtime, pred_w_900_4.pred_runtime_user)
mae_heu_sa4 = mean_absolute_error(pred_w_900_4.gt_runtime, pred_w_900_4.pred_runtime_heuristic)
mae_dt_sa4  = mean_absolute_error(pred_w_900_4.gt_runtime, pred_w_900_4.pred_runtime_dt_fs2)
mae_rnp_sa4 = mean_absolute_error(pred_w_900_4.gt_runtime, pred_w_900_4.pred_runtime_rnp_fs2)
mae_knn_sa4 = mean_absolute_error(pred_w_900_4.gt_runtime, pred_w_900_4.pred_runtime_knn_fs2)
mae_llm_sa4 = mean_absolute_error(pred_w_900_4.gt_runtime, pred_w_900_4.pred_runtime_ora)

mae_usr_sa5 = mean_absolute_error(pred_w_900_5.gt_runtime, pred_w_900_5.pred_runtime_user)
mae_heu_sa5 = mean_absolute_error(pred_w_900_5.gt_runtime, pred_w_900_5.pred_runtime_heuristic)
mae_dt_sa5  = mean_absolute_error(pred_w_900_5.gt_runtime, pred_w_900_5.pred_runtime_dt_fs2)
mae_rnp_sa5 = mean_absolute_error(pred_w_900_5.gt_runtime, pred_w_900_5.pred_runtime_rnp_fs2)
mae_knn_sa5 = mean_absolute_error(pred_w_900_5.gt_runtime, pred_w_900_5.pred_runtime_knn_fs2)
mae_llm_sa5 = mean_absolute_error(pred_w_900_5.gt_runtime, pred_w_900_5.pred_runtime_ora)

ea_usr_sa1 = estimation_accuracy(pred_w_900_1.gt_runtime, pred_w_900_1.pred_runtime_user)
ea_heu_sa1 = estimation_accuracy(pred_w_900_1.gt_runtime, pred_w_900_1.pred_runtime_heuristic)
ea_dt_sa1  = estimation_accuracy(pred_w_900_1.gt_runtime, pred_w_900_1.pred_runtime_dt_fs2)
ea_rnp_sa1 = estimation_accuracy(pred_w_900_1.gt_runtime, pred_w_900_1.pred_runtime_rnp_fs2)
ea_knn_sa1 = estimation_accuracy(pred_w_900_1.gt_runtime, pred_w_900_1.pred_runtime_knn_fs2)
ea_llm_sa1 = estimation_accuracy(pred_w_900_1.gt_runtime, pred_w_900_1.pred_runtime_ora)

ea_usr_sa2 = estimation_accuracy(pred_w_900_2.gt_runtime, pred_w_900_2.pred_runtime_user)
ea_heu_sa2 = estimation_accuracy(pred_w_900_2.gt_runtime, pred_w_900_2.pred_runtime_heuristic)
ea_dt_sa2  = estimation_accuracy(pred_w_900_2.gt_runtime, pred_w_900_2.pred_runtime_dt_fs2)
ea_rnp_sa2 = estimation_accuracy(pred_w_900_2.gt_runtime, pred_w_900_2.pred_runtime_rnp_fs2)
ea_knn_sa2 = estimation_accuracy(pred_w_900_2.gt_runtime, pred_w_900_2.pred_runtime_knn_fs2)
ea_llm_sa2 = estimation_accuracy(pred_w_900_2.gt_runtime, pred_w_900_2.pred_runtime_ora)

ea_usr_sa3 = estimation_accuracy(pred_w_900_3.gt_runtime, pred_w_900_3.pred_runtime_user)
ea_heu_sa3 = estimation_accuracy(pred_w_900_3.gt_runtime, pred_w_900_3.pred_runtime_heuristic)
ea_dt_sa3  = estimation_accuracy(pred_w_900_3.gt_runtime, pred_w_900_3.pred_runtime_dt_fs2)
ea_rnp_sa3 = estimation_accuracy(pred_w_900_3.gt_runtime, pred_w_900_3.pred_runtime_rnp_fs2)
ea_knn_sa3 = estimation_accuracy(pred_w_900_3.gt_runtime, pred_w_900_3.pred_runtime_knn_fs2)
ea_llm_sa3 = estimation_accuracy(pred_w_900_3.gt_runtime, pred_w_900_3.pred_runtime_ora)

ea_usr_sa4 = estimation_accuracy(pred_w_900_4.gt_runtime, pred_w_900_4.pred_runtime_user)
ea_heu_sa4 = estimation_accuracy(pred_w_900_4.gt_runtime, pred_w_900_4.pred_runtime_heuristic)
ea_dt_sa4  = estimation_accuracy(pred_w_900_4.gt_runtime, pred_w_900_4.pred_runtime_dt_fs2)
ea_rnp_sa4 = estimation_accuracy(pred_w_900_4.gt_runtime, pred_w_900_4.pred_runtime_rnp_fs2)
ea_knn_sa4 = estimation_accuracy(pred_w_900_4.gt_runtime, pred_w_900_4.pred_runtime_knn_fs2)
ea_llm_sa4 = estimation_accuracy(pred_w_900_4.gt_runtime, pred_w_900_4.pred_runtime_ora)

ea_usr_sa5 = estimation_accuracy(pred_w_900_5.gt_runtime, pred_w_900_5.pred_runtime_user)
ea_heu_sa5 = estimation_accuracy(pred_w_900_5.gt_runtime, pred_w_900_5.pred_runtime_heuristic)
ea_dt_sa5  = estimation_accuracy(pred_w_900_5.gt_runtime, pred_w_900_5.pred_runtime_dt_fs2)
ea_rnp_sa5 = estimation_accuracy(pred_w_900_5.gt_runtime, pred_w_900_5.pred_runtime_rnp_fs2)
ea_knn_sa5 = estimation_accuracy(pred_w_900_5.gt_runtime, pred_w_900_5.pred_runtime_knn_fs2)
ea_llm_sa5 = estimation_accuracy(pred_w_900_5.gt_runtime, pred_w_900_5.pred_runtime_ora)

mape_usr_sa1 = mean_average_percentage_error(pred_w_900_1.gt_runtime, pred_w_900_1.pred_runtime_user)
mape_heu_sa1 = mean_average_percentage_error(pred_w_900_1.gt_runtime, pred_w_900_1.pred_runtime_heuristic)
mape_dt_sa1  = mean_average_percentage_error(pred_w_900_1.gt_runtime, pred_w_900_1.pred_runtime_dt_fs2)
mape_rnp_sa1 = mean_average_percentage_error(pred_w_900_1.gt_runtime, pred_w_900_1.pred_runtime_rnp_fs2)
mape_knn_sa1 = mean_average_percentage_error(pred_w_900_1.gt_runtime, pred_w_900_1.pred_runtime_knn_fs2)
mape_llm_sa1 = mean_average_percentage_error(pred_w_900_1.gt_runtime, pred_w_900_1.pred_runtime_ora)

mape_usr_sa2 = mean_average_percentage_error(pred_w_900_2.gt_runtime, pred_w_900_2.pred_runtime_user)
mape_heu_sa2 = mean_average_percentage_error(pred_w_900_2.gt_runtime, pred_w_900_2.pred_runtime_heuristic)
mape_dt_sa2  = mean_average_percentage_error(pred_w_900_2.gt_runtime, pred_w_900_2.pred_runtime_dt_fs2)
mape_rnp_sa2 = mean_average_percentage_error(pred_w_900_2.gt_runtime, pred_w_900_2.pred_runtime_rnp_fs2)
mape_knn_sa2 = mean_average_percentage_error(pred_w_900_2.gt_runtime, pred_w_900_2.pred_runtime_knn_fs2)
mape_llm_sa2 = mean_average_percentage_error(pred_w_900_2.gt_runtime, pred_w_900_2.pred_runtime_ora)

mape_usr_sa3 = mean_average_percentage_error(pred_w_900_3.gt_runtime, pred_w_900_3.pred_runtime_user)
mape_heu_sa3 = mean_average_percentage_error(pred_w_900_3.gt_runtime, pred_w_900_3.pred_runtime_heuristic)
mape_dt_sa3  = mean_average_percentage_error(pred_w_900_3.gt_runtime, pred_w_900_3.pred_runtime_dt_fs2)
mape_rnp_sa3 = mean_average_percentage_error(pred_w_900_3.gt_runtime, pred_w_900_3.pred_runtime_rnp_fs2)
mape_knn_sa3 = mean_average_percentage_error(pred_w_900_3.gt_runtime, pred_w_900_3.pred_runtime_knn_fs2)
mape_llm_sa3 = mean_average_percentage_error(pred_w_900_3.gt_runtime, pred_w_900_3.pred_runtime_ora)

mape_usr_sa4 = mean_average_percentage_error(pred_w_900_4.gt_runtime, pred_w_900_4.pred_runtime_user)
mape_heu_sa4 = mean_average_percentage_error(pred_w_900_4.gt_runtime, pred_w_900_4.pred_runtime_heuristic)
mape_dt_sa4  = mean_average_percentage_error(pred_w_900_4.gt_runtime, pred_w_900_4.pred_runtime_dt_fs2)
mape_rnp_sa4 = mean_average_percentage_error(pred_w_900_4.gt_runtime, pred_w_900_4.pred_runtime_rnp_fs2)
mape_knn_sa4 = mean_average_percentage_error(pred_w_900_4.gt_runtime, pred_w_900_4.pred_runtime_knn_fs2)
mape_llm_sa4 = mean_average_percentage_error(pred_w_900_4.gt_runtime, pred_w_900_4.pred_runtime_ora)

mape_usr_sa5 = mean_average_percentage_error(pred_w_900_5.gt_runtime, pred_w_900_5.pred_runtime_user)
mape_heu_sa5 = mean_average_percentage_error(pred_w_900_5.gt_runtime, pred_w_900_5.pred_runtime_heuristic)
mape_dt_sa5  = mean_average_percentage_error(pred_w_900_5.gt_runtime, pred_w_900_5.pred_runtime_dt_fs2)
mape_rnp_sa5 = mean_average_percentage_error(pred_w_900_5.gt_runtime, pred_w_900_5.pred_runtime_rnp_fs2)
mape_knn_sa5 = mean_average_percentage_error(pred_w_900_5.gt_runtime, pred_w_900_5.pred_runtime_knn_fs2)
mape_llm_sa5 = mean_average_percentage_error(pred_w_900_5.gt_runtime, pred_w_900_5.pred_runtime_ora)

mae_usr_avg = (mae_usr_sa1 + mae_usr_sa2 + mae_usr_sa3 + mae_usr_sa4 + mae_usr_sa5)/5
mae_heu_avg = (mae_heu_sa1 + mae_heu_sa2 + mae_heu_sa3 + mae_heu_sa4 + mae_heu_sa5)/5
mae_dt_avg  = (mae_dt_sa1 + mae_dt_sa2 + mae_dt_sa3 + mae_dt_sa4 + mae_dt_sa5)/5
mae_rnp_avg = (mae_rnp_sa1 + mae_rnp_sa2 + mae_rnp_sa3 + mae_rnp_sa4 + mae_rnp_sa5)/5 
mae_knn_avg = (mae_knn_sa1 + mae_knn_sa2 + mae_knn_sa3 + mae_knn_sa4 + mae_knn_sa5)/5
mae_llm_avg = (mae_llm_sa1 + mae_llm_sa2 + mae_llm_sa3 + mae_llm_sa4 + mae_llm_sa5)/5 

ea_usr_avg = (ea_usr_sa1 + ea_usr_sa2 + ea_usr_sa3 + ea_usr_sa4 + ea_usr_sa5)/5
ea_heu_avg = (ea_heu_sa1 + ea_heu_sa2 + ea_heu_sa3 + ea_heu_sa4 + ea_heu_sa5)/5
ea_dt_avg  = (ea_dt_sa1 + ea_dt_sa2 + ea_dt_sa3 + ea_dt_sa4 + ea_dt_sa5)/5
ea_rnp_avg = (ea_rnp_sa1 + ea_rnp_sa2 + ea_rnp_sa3 + ea_rnp_sa4 + ea_rnp_sa5)/5
ea_knn_avg = (ea_knn_sa1 + ea_knn_sa2 + ea_knn_sa3 + ea_knn_sa4 + ea_knn_sa5)/5
ea_llm_avg = (ea_llm_sa1 + ea_llm_sa2 + ea_llm_sa3 + ea_llm_sa4 + ea_llm_sa5)/5

mape_usr_avg = (mape_usr_sa1 + mape_usr_sa2 + mape_usr_sa3 + mape_usr_sa4 + mape_usr_sa5)/5
mape_heu_avg = (mape_heu_sa1 + mape_heu_sa2 + mape_heu_sa3 + mape_heu_sa4 + mape_heu_sa5)/5
mape_dt_avg  = (mape_dt_sa1 + mape_dt_sa2 + mape_dt_sa3 + mape_dt_sa4 + mape_dt_sa5)/5
mape_rnp_avg = (mape_rnp_sa1 + mape_rnp_sa2 + mape_rnp_sa3 + mape_rnp_sa4 + mape_rnp_sa5)/5
mape_knn_avg = (mape_knn_sa1 + mape_knn_sa2 + mape_knn_sa3 + mape_knn_sa4 + mape_knn_sa5)/5
mape_llm_avg = (mape_llm_sa1 + mape_llm_sa2 + mape_llm_sa3 + mape_llm_sa4 + mape_llm_sa5)/5

### Average Metrics

res_avg = pd.DataFrame()
res_avg['Workload'] = ('A', 'A', 'A', 'A', 'A', 'A')
res_avg['Method'] = ('User', 'Heuristic', 'DT', 'RNP', 'KNN', 'ORA')
res_avg['MAE'] = (mae_usr_avg, mae_heu_avg, mae_dt_avg, mae_rnp_avg, mae_knn_avg, mae_llm_avg)
res_avg['Average Error'] = (convert_to_hms(mae_usr_avg), convert_to_hms(mae_heu_avg), convert_to_hms(mae_dt_avg), convert_to_hms(mae_rnp_avg), convert_to_hms(mae_knn_avg), convert_to_hms(mae_llm_avg))
res_avg['EA'] = (ea_usr_avg, ea_heu_avg, ea_dt_avg, ea_rnp_avg, ea_knn_avg, ea_llm_avg)
res_avg['MAPE'] = (mape_usr_avg/1000, mape_heu_avg/1000, mape_dt_avg/1000, mape_rnp_avg/1000, mape_knn_avg/1000, mape_llm_avg/1000)

res_avg

### Details

res_sa1 = pd.DataFrame()
res_sa1['Workload'] = ('H1', 'H1', 'H1', 'H1', 'H1', 'H1')
res_sa1['Method'] = ('User', 'Heuristic', 'DT', 'RNP', 'KNN', 'ORA')
res_sa1['MAE'] = (mae_usr_sa1, mae_heu_sa1, mae_dt_sa1, mae_rnp_sa1, mae_knn_sa1, mae_llm_sa1)
res_sa1['Average Error'] = (convert_to_hms(mae_usr_sa1), convert_to_hms(mae_heu_sa1), convert_to_hms(mae_dt_sa1), convert_to_hms(mae_rnp_sa1), convert_to_hms(mae_knn_sa1), convert_to_hms(mae_llm_sa1))
res_sa1['EA'] = (ea_usr_sa1, ea_heu_sa1, ea_dt_sa1, ea_rnp_sa1, ea_knn_sa1, ea_llm_sa1)
res_sa1['MAPE'] = (mape_usr_sa1/1000, mape_heu_sa1/1000, mape_dt_sa1/1000, mape_rnp_sa1/1000, mape_knn_sa1/1000, mape_llm_sa1/1000)

res_sa1

res_sa2 = pd.DataFrame()
res_sa2['Workload'] = ('H2', 'H2', 'H2', 'H2', 'H2', 'H2')
res_sa2['Method'] = ('User', 'Heuristic', 'DT', 'RNP', 'KNN', 'ORA')
res_sa2['MAE'] = (mae_usr_sa2, mae_heu_sa2, mae_dt_sa2, mae_rnp_sa2, mae_knn_sa2, mae_llm_sa2)
res_sa2['Average Error'] = (convert_to_hms(mae_usr_sa2), convert_to_hms(mae_heu_sa2), convert_to_hms(mae_dt_sa2), convert_to_hms(mae_rnp_sa2), convert_to_hms(mae_knn_sa2), convert_to_hms(mae_llm_sa2))
res_sa2['EA'] = (ea_usr_sa2, ea_heu_sa2, ea_dt_sa2, ea_rnp_sa2, ea_knn_sa2, ea_llm_sa2)
res_sa2['MAPE'] = (mape_usr_sa2/1000, mape_heu_sa2/1000, mape_dt_sa2/1000, mape_rnp_sa2/1000, mape_knn_sa2/1000, mape_llm_sa2/1000)

res_sa2

res_sa3 = pd.DataFrame()
res_sa3['Workload'] = ('H3', 'H3', 'H3', 'H3', 'H3', 'H3')
res_sa3['Method'] = ('User', 'Heuristic', 'DT', 'RNP', 'KNN', 'ORA')
res_sa3['MAE'] = (mae_usr_sa3, mae_heu_sa3, mae_dt_sa3, mae_rnp_sa3, mae_knn_sa3, mae_llm_sa3)
res_sa3['Average Error'] = (convert_to_hms(mae_usr_sa3), convert_to_hms(mae_heu_sa3), convert_to_hms(mae_dt_sa3), convert_to_hms(mae_rnp_sa3), convert_to_hms(mae_knn_sa3), convert_to_hms(mae_llm_sa3))
res_sa3['EA'] = (ea_usr_sa3, ea_heu_sa3, ea_dt_sa3, ea_rnp_sa3, ea_knn_sa3, ea_llm_sa3)
res_sa3['MAPE'] = (mape_usr_sa3/1000, mape_heu_sa3/1000, mape_dt_sa3/1000, mape_rnp_sa3/1000, mape_knn_sa3/1000, mape_llm_sa3/1000)

res_sa3

res_sa4 = pd.DataFrame()
res_sa4['Workload'] = ('H4', 'H4', 'H4', 'H4', 'H4', 'H4')
res_sa4['Method'] = ('User', 'Heuristic', 'DT', 'RNP', 'KNN', 'ORA')
res_sa4['MAE'] = (mae_usr_sa4, mae_heu_sa4, mae_dt_sa4, mae_rnp_sa4, mae_knn_sa4, mae_llm_sa4)
res_sa4['Average Error'] = (convert_to_hms(mae_usr_sa4), convert_to_hms(mae_heu_sa4), convert_to_hms(mae_dt_sa4), convert_to_hms(mae_rnp_sa4), convert_to_hms(mae_knn_sa4), convert_to_hms(mae_llm_sa4))
res_sa4['EA'] = (ea_usr_sa4, ea_heu_sa4, ea_dt_sa4, ea_rnp_sa4, ea_knn_sa4, ea_llm_sa4)
res_sa4['MAPE'] = (mape_usr_sa4/1000, mape_heu_sa4/1000, mape_dt_sa4/1000, mape_rnp_sa4/1000, mape_knn_sa4/1000, mape_llm_sa4/1000)

res_sa4

res_sa5 = pd.DataFrame()
res_sa5['Workload'] = ('H5', 'H5', 'H5', 'H5', 'H5', 'H5')
res_sa5['Method'] = ('User', 'Heuristic', 'DT', 'RNP', 'KNN', 'ORA')
res_sa5['MAE'] = (mae_usr_sa5, mae_heu_sa5, mae_dt_sa5, mae_rnp_sa5, mae_knn_sa5, mae_llm_sa5)
res_sa5['Average Error'] = (convert_to_hms(mae_usr_sa5), convert_to_hms(mae_heu_sa5), convert_to_hms(mae_dt_sa5), convert_to_hms(mae_rnp_sa5), convert_to_hms(mae_knn_sa5), convert_to_hms(mae_llm_sa5))
res_sa5['EA'] = (ea_usr_sa5, ea_heu_sa5, ea_dt_sa5, ea_rnp_sa5, ea_knn_sa5, ea_llm_sa5)
res_sa5['MAPE'] = (mape_usr_sa5/1000, mape_heu_sa5/1000, mape_dt_sa5/1000, mape_rnp_sa5/1000, mape_knn_sa5/1000, mape_llm_sa5/1000)

res_sa5

results = pd.concat([res, res_avg, res_sa1, res_sa2, res_sa3, res_sa4, res_sa5], axis=0)
results = results.reset_index(drop=True)
results

results.to_csv("metrics/pm100_900_metrics.csv", index=False)

## Scenario (b): 9000x980

### Metric Computation

mae_usr_sb1 = mean_absolute_error(pred_w_9k_1.gt_runtime, pred_w_9k_1.pred_runtime_user)
mae_heu_sb1 = mean_absolute_error(pred_w_9k_1.gt_runtime, pred_w_9k_1.pred_runtime_heuristic)
mae_dt_sb1  = mean_absolute_error(pred_w_9k_1.gt_runtime, pred_w_9k_1.pred_runtime_dt_fs2)
mae_rnp_sb1 = mean_absolute_error(pred_w_9k_1.gt_runtime, pred_w_9k_1.pred_runtime_rnp_fs2)
mae_knn_sb1 = mean_absolute_error(pred_w_9k_1.gt_runtime, pred_w_9k_1.pred_runtime_knn_fs2)
mae_llm_sb1 = mean_absolute_error(pred_w_9k_1.gt_runtime, pred_w_9k_1.pred_runtime_ora)

mae_usr_sb2 = mean_absolute_error(pred_w_9k_2.gt_runtime, pred_w_9k_2.pred_runtime_user)
mae_heu_sb2 = mean_absolute_error(pred_w_9k_2.gt_runtime, pred_w_9k_2.pred_runtime_heuristic)
mae_dt_sb2  = mean_absolute_error(pred_w_9k_2.gt_runtime, pred_w_9k_2.pred_runtime_dt_fs2)
mae_rnp_sb2 = mean_absolute_error(pred_w_9k_2.gt_runtime, pred_w_9k_2.pred_runtime_rnp_fs2)
mae_knn_sb2 = mean_absolute_error(pred_w_9k_2.gt_runtime, pred_w_9k_2.pred_runtime_knn_fs2)
mae_llm_sb2 = mean_absolute_error(pred_w_9k_2.gt_runtime, pred_w_9k_2.pred_runtime_ora)

mae_usr_sb3 = mean_absolute_error(pred_w_9k_3.gt_runtime, pred_w_9k_3.pred_runtime_user)
mae_heu_sb3 = mean_absolute_error(pred_w_9k_3.gt_runtime, pred_w_9k_3.pred_runtime_heuristic)
mae_dt_sb3  = mean_absolute_error(pred_w_9k_3.gt_runtime, pred_w_9k_3.pred_runtime_dt_fs2)
mae_rnp_sb3 = mean_absolute_error(pred_w_9k_3.gt_runtime, pred_w_9k_3.pred_runtime_rnp_fs2)
mae_knn_sb3 = mean_absolute_error(pred_w_9k_3.gt_runtime, pred_w_9k_3.pred_runtime_knn_fs2)
mae_llm_sb3 = mean_absolute_error(pred_w_9k_3.gt_runtime, pred_w_9k_3.pred_runtime_ora)

mae_usr_sb4 = mean_absolute_error(pred_w_9k_4.gt_runtime, pred_w_9k_4.pred_runtime_user)
mae_heu_sb4 = mean_absolute_error(pred_w_9k_4.gt_runtime, pred_w_9k_4.pred_runtime_heuristic)
mae_dt_sb4  = mean_absolute_error(pred_w_9k_4.gt_runtime, pred_w_9k_4.pred_runtime_dt_fs2)
mae_rnp_sb4 = mean_absolute_error(pred_w_9k_4.gt_runtime, pred_w_9k_4.pred_runtime_rnp_fs2)
mae_knn_sb4 = mean_absolute_error(pred_w_9k_4.gt_runtime, pred_w_9k_4.pred_runtime_knn_fs2)
mae_llm_sb4 = mean_absolute_error(pred_w_9k_4.gt_runtime, pred_w_9k_4.pred_runtime_ora)

mae_usr_sb5 = mean_absolute_error(pred_w_9k_5.gt_runtime, pred_w_9k_5.pred_runtime_user)
mae_heu_sb5 = mean_absolute_error(pred_w_9k_5.gt_runtime, pred_w_9k_5.pred_runtime_heuristic)
mae_dt_sb5  = mean_absolute_error(pred_w_9k_5.gt_runtime, pred_w_9k_5.pred_runtime_dt_fs2)
mae_rnp_sb5 = mean_absolute_error(pred_w_9k_5.gt_runtime, pred_w_9k_5.pred_runtime_rnp_fs2)
mae_knn_sb5 = mean_absolute_error(pred_w_9k_5.gt_runtime, pred_w_9k_5.pred_runtime_knn_fs2)
mae_llm_sb5 = mean_absolute_error(pred_w_9k_5.gt_runtime, pred_w_9k_5.pred_runtime_ora)

ea_usr_sb1 = estimation_accuracy(pred_w_9k_1.gt_runtime, pred_w_9k_1.pred_runtime_user)
ea_heu_sb1 = estimation_accuracy(pred_w_9k_1.gt_runtime, pred_w_9k_1.pred_runtime_heuristic)
ea_dt_sb1  = estimation_accuracy(pred_w_9k_1.gt_runtime, pred_w_9k_1.pred_runtime_dt_fs2)
ea_rnp_sb1 = estimation_accuracy(pred_w_9k_1.gt_runtime, pred_w_9k_1.pred_runtime_rnp_fs2)
ea_knn_sb1 = estimation_accuracy(pred_w_9k_1.gt_runtime, pred_w_9k_1.pred_runtime_knn_fs2)
ea_llm_sb1 = estimation_accuracy(pred_w_9k_1.gt_runtime, pred_w_9k_1.pred_runtime_ora)

ea_usr_sb2 = estimation_accuracy(pred_w_9k_2.gt_runtime, pred_w_9k_2.pred_runtime_user)
ea_heu_sb2 = estimation_accuracy(pred_w_9k_2.gt_runtime, pred_w_9k_2.pred_runtime_heuristic)
ea_dt_sb2  = estimation_accuracy(pred_w_9k_2.gt_runtime, pred_w_9k_2.pred_runtime_dt_fs2)
ea_rnp_sb2 = estimation_accuracy(pred_w_9k_2.gt_runtime, pred_w_9k_2.pred_runtime_rnp_fs2)
ea_knn_sb2 = estimation_accuracy(pred_w_9k_2.gt_runtime, pred_w_9k_2.pred_runtime_knn_fs2)
ea_llm_sb2 = estimation_accuracy(pred_w_9k_2.gt_runtime, pred_w_9k_2.pred_runtime_ora)

ea_usr_sb3 = estimation_accuracy(pred_w_9k_3.gt_runtime, pred_w_9k_3.pred_runtime_user)
ea_heu_sb3 = estimation_accuracy(pred_w_9k_3.gt_runtime, pred_w_9k_3.pred_runtime_heuristic)
ea_dt_sb3  = estimation_accuracy(pred_w_9k_3.gt_runtime, pred_w_9k_3.pred_runtime_dt_fs2)
ea_rnp_sb3 = estimation_accuracy(pred_w_9k_3.gt_runtime, pred_w_9k_3.pred_runtime_rnp_fs2)
ea_knn_sb3 = estimation_accuracy(pred_w_9k_3.gt_runtime, pred_w_9k_3.pred_runtime_knn_fs2)
ea_llm_sb3 = estimation_accuracy(pred_w_9k_3.gt_runtime, pred_w_9k_3.pred_runtime_ora)

ea_usr_sb4 = estimation_accuracy(pred_w_9k_4.gt_runtime, pred_w_9k_4.pred_runtime_user)
ea_heu_sb4 = estimation_accuracy(pred_w_9k_4.gt_runtime, pred_w_9k_4.pred_runtime_heuristic)
ea_dt_sb4  = estimation_accuracy(pred_w_9k_4.gt_runtime, pred_w_9k_4.pred_runtime_dt_fs2)
ea_rnp_sb4 = estimation_accuracy(pred_w_9k_4.gt_runtime, pred_w_9k_4.pred_runtime_rnp_fs2)
ea_knn_sb4 = estimation_accuracy(pred_w_9k_4.gt_runtime, pred_w_9k_4.pred_runtime_knn_fs2)
ea_llm_sb4 = estimation_accuracy(pred_w_9k_4.gt_runtime, pred_w_9k_4.pred_runtime_ora)

ea_usr_sb5 = estimation_accuracy(pred_w_9k_5.gt_runtime, pred_w_9k_5.pred_runtime_user)
ea_heu_sb5 = estimation_accuracy(pred_w_9k_5.gt_runtime, pred_w_9k_5.pred_runtime_heuristic)
ea_dt_sb5  = estimation_accuracy(pred_w_9k_5.gt_runtime, pred_w_9k_5.pred_runtime_dt_fs2)
ea_rnp_sb5 = estimation_accuracy(pred_w_9k_5.gt_runtime, pred_w_9k_5.pred_runtime_rnp_fs2)
ea_knn_sb5 = estimation_accuracy(pred_w_9k_5.gt_runtime, pred_w_9k_5.pred_runtime_knn_fs2)
ea_llm_sb5 = estimation_accuracy(pred_w_9k_5.gt_runtime, pred_w_9k_5.pred_runtime_ora)

mape_usr_sb1 = mean_average_percentage_error(pred_w_9k_1.gt_runtime, pred_w_9k_1.pred_runtime_user)
mape_heu_sb1 = mean_average_percentage_error(pred_w_9k_1.gt_runtime, pred_w_9k_1.pred_runtime_heuristic)
mape_dt_sb1  = mean_average_percentage_error(pred_w_9k_1.gt_runtime, pred_w_9k_1.pred_runtime_dt_fs2)
mape_rnp_sb1 = mean_average_percentage_error(pred_w_9k_1.gt_runtime, pred_w_9k_1.pred_runtime_rnp_fs2)
mape_knn_sb1 = mean_average_percentage_error(pred_w_9k_1.gt_runtime, pred_w_9k_1.pred_runtime_knn_fs2)
mape_llm_sb1 = mean_average_percentage_error(pred_w_9k_1.gt_runtime, pred_w_9k_1.pred_runtime_ora)

mape_usr_sb2 = mean_average_percentage_error(pred_w_9k_2.gt_runtime, pred_w_9k_2.pred_runtime_user)
mape_heu_sb2 = mean_average_percentage_error(pred_w_9k_2.gt_runtime, pred_w_9k_2.pred_runtime_heuristic)
mape_dt_sb2  = mean_average_percentage_error(pred_w_9k_2.gt_runtime, pred_w_9k_2.pred_runtime_dt_fs2)
mape_rnp_sb2 = mean_average_percentage_error(pred_w_9k_2.gt_runtime, pred_w_9k_2.pred_runtime_rnp_fs2)
mape_knn_sb2 = mean_average_percentage_error(pred_w_9k_2.gt_runtime, pred_w_9k_2.pred_runtime_knn_fs2)
mape_llm_sb2 = mean_average_percentage_error(pred_w_9k_2.gt_runtime, pred_w_9k_2.pred_runtime_ora)

mape_usr_sb3 = mean_average_percentage_error(pred_w_9k_3.gt_runtime, pred_w_9k_3.pred_runtime_user)
mape_heu_sb3 = mean_average_percentage_error(pred_w_9k_3.gt_runtime, pred_w_9k_3.pred_runtime_heuristic)
mape_dt_sb3  = mean_average_percentage_error(pred_w_9k_3.gt_runtime, pred_w_9k_3.pred_runtime_dt_fs2)
mape_rnp_sb3 = mean_average_percentage_error(pred_w_9k_3.gt_runtime, pred_w_9k_3.pred_runtime_rnp_fs2)
mape_knn_sb3 = mean_average_percentage_error(pred_w_9k_3.gt_runtime, pred_w_9k_3.pred_runtime_knn_fs2)
mape_llm_sb3 = mean_average_percentage_error(pred_w_9k_3.gt_runtime, pred_w_9k_3.pred_runtime_ora)

mape_usr_sb4 = mean_average_percentage_error(pred_w_9k_4.gt_runtime, pred_w_9k_4.pred_runtime_user)
mape_heu_sb4 = mean_average_percentage_error(pred_w_9k_4.gt_runtime, pred_w_9k_4.pred_runtime_heuristic)
mape_dt_sb4  = mean_average_percentage_error(pred_w_9k_4.gt_runtime, pred_w_9k_4.pred_runtime_dt_fs2)
mape_rnp_sb4 = mean_average_percentage_error(pred_w_9k_4.gt_runtime, pred_w_9k_4.pred_runtime_rnp_fs2)
mape_knn_sb4 = mean_average_percentage_error(pred_w_9k_4.gt_runtime, pred_w_9k_4.pred_runtime_knn_fs2)
mape_llm_sb4 = mean_average_percentage_error(pred_w_9k_4.gt_runtime, pred_w_9k_4.pred_runtime_ora)

mape_usr_sb5 = mean_average_percentage_error(pred_w_9k_5.gt_runtime, pred_w_9k_5.pred_runtime_user)
mape_heu_sb5 = mean_average_percentage_error(pred_w_9k_5.gt_runtime, pred_w_9k_5.pred_runtime_heuristic)
mape_dt_sb5  = mean_average_percentage_error(pred_w_9k_5.gt_runtime, pred_w_9k_5.pred_runtime_dt_fs2)
mape_rnp_sb5 = mean_average_percentage_error(pred_w_9k_5.gt_runtime, pred_w_9k_5.pred_runtime_rnp_fs2)
mape_knn_sb5 = mean_average_percentage_error(pred_w_9k_5.gt_runtime, pred_w_9k_5.pred_runtime_knn_fs2)
mape_llm_sb5 = mean_average_percentage_error(pred_w_9k_5.gt_runtime, pred_w_9k_5.pred_runtime_ora)

mae_usr_avg = (mae_usr_sb1 + mae_usr_sb2 + mae_usr_sb3 + mae_usr_sb4 + mae_usr_sb5)/5
mae_heu_avg = (mae_heu_sb1 + mae_heu_sb2 + mae_heu_sb3 + mae_heu_sb4 + mae_heu_sb5)/5
mae_dt_avg  = (mae_dt_sb1 + mae_dt_sb2 + mae_dt_sb3 + mae_dt_sb4 + mae_dt_sb5)/5
mae_rnp_avg = (mae_rnp_sb1 + mae_rnp_sb2 + mae_rnp_sb3 + mae_rnp_sb4 + mae_rnp_sb5)/5 
mae_knn_avg = (mae_knn_sb1 + mae_knn_sb2 + mae_knn_sb3 + mae_knn_sb4 + mae_knn_sb5)/5
mae_llm_avg = (mae_llm_sb1 + mae_llm_sb2 + mae_llm_sb3 + mae_llm_sb4 + mae_llm_sb5)/5 

ea_usr_avg = (ea_usr_sb1 + ea_usr_sb2 + ea_usr_sb3 + ea_usr_sb4 + ea_usr_sb5)/5
ea_heu_avg = (ea_heu_sb1 + ea_heu_sb2 + ea_heu_sb3 + ea_heu_sb4 + ea_heu_sb5)/5
ea_dt_avg  = (ea_dt_sb1 + ea_dt_sb2 + ea_dt_sb3 + ea_dt_sb4 + ea_dt_sb5)/5
ea_rnp_avg = (ea_rnp_sb1 + ea_rnp_sb2 + ea_rnp_sb3 + ea_rnp_sb4 + ea_rnp_sb5)/5
ea_knn_avg = (ea_knn_sb1 + ea_knn_sb2 + ea_knn_sb3 + ea_knn_sb4 + ea_knn_sb5)/5
ea_llm_avg = (ea_llm_sb1 + ea_llm_sb2 + ea_llm_sb3 + ea_llm_sb4 + ea_llm_sb5)/5

mape_usr_avg = (mape_usr_sb1 + mape_usr_sb2 + mape_usr_sb3 + mape_usr_sb4 + mape_usr_sb5)/5
mape_heu_avg = (mape_heu_sb1 + mape_heu_sb2 + mape_heu_sb3 + mape_heu_sb4 + mape_heu_sb5)/5
mape_dt_avg  = (mape_dt_sb1 + mape_dt_sb2 + mape_dt_sb3 + mape_dt_sb4 + mape_dt_sb5)/5
mape_rnp_avg = (mape_rnp_sb1 + mape_rnp_sb2 + mape_rnp_sb3 + mape_rnp_sb4 + mape_rnp_sb5)/5
mape_knn_avg = (mape_knn_sb1 + mape_knn_sb2 + mape_knn_sb3 + mape_knn_sb4 + mape_knn_sb5)/5
mape_llm_avg = (mape_llm_sb1 + mape_llm_sb2 + mape_llm_sb3 + mape_llm_sb4 + mape_llm_sb5)/5

### Average Metrics

res_avg = pd.DataFrame()
res_avg['Workload'] = ('A', 'A', 'A', 'A', 'A', 'A')
res_avg['Method'] = ('User', 'Heuristic', 'DT', 'RNP', 'KNN', 'ORA')
res_avg['MAE'] = (mae_usr_avg, mae_heu_avg, mae_dt_avg, mae_rnp_avg, mae_knn_avg, mae_llm_avg)
res_avg['Average Error'] = (convert_to_hms(mae_usr_avg), convert_to_hms(mae_heu_avg), convert_to_hms(mae_dt_avg), convert_to_hms(mae_rnp_avg), convert_to_hms(mae_knn_avg), convert_to_hms(mae_llm_avg))
res_avg['EA'] = (ea_usr_avg, ea_heu_avg, ea_dt_avg, ea_rnp_avg, ea_knn_avg, ea_llm_avg)
res_avg['MAPE'] = (mape_usr_avg/1000, mape_heu_avg/1000, mape_dt_avg/1000, mape_rnp_avg/1000, mape_knn_avg/1000, mape_llm_avg/1000)

res_avg

### Details

res_sb1 = pd.DataFrame()
res_sb1['Workload'] = ('H1', 'H1', 'H1', 'H1', 'H1', 'H1')
res_sb1['Method'] = ('User', 'Heuristic', 'DT', 'RNP', 'KNN', 'ORA')
res_sb1['MAE'] = (mae_usr_sb1, mae_heu_sb1, mae_dt_sb1, mae_rnp_sb1, mae_knn_sb1, mae_llm_sb1)
res_sb1['Average Error'] = (convert_to_hms(mae_usr_sb1), convert_to_hms(mae_heu_sb1), convert_to_hms(mae_dt_sb1), convert_to_hms(mae_rnp_sb1), convert_to_hms(mae_knn_sb1), convert_to_hms(mae_llm_sb1))
res_sb1['EA'] = (ea_usr_sb1, ea_heu_sb1, ea_dt_sb1, ea_rnp_sb1, ea_knn_sb1, ea_llm_sb1)
res_sb1['MAPE'] = (mape_usr_sb1/1000, mape_heu_sb1/1000, mape_dt_sb1/1000, mape_rnp_sb1/1000, mape_knn_sb1/1000, mape_llm_sb1/1000)

res_sb1

res_sb2 = pd.DataFrame()
res_sb2['Workload'] = ('H2', 'H2', 'H2', 'H2', 'H2', 'H2')
res_sb2['Method'] = ('User', 'Heuristic', 'DT', 'RNP', 'KNN', 'ORA')
res_sb2['MAE'] = (mae_usr_sb2, mae_heu_sb2, mae_dt_sb2, mae_rnp_sb2, mae_knn_sb2, mae_llm_sb2)
res_sb2['Average Error'] = (convert_to_hms(mae_usr_sb2), convert_to_hms(mae_heu_sb2), convert_to_hms(mae_dt_sb2), convert_to_hms(mae_rnp_sb2), convert_to_hms(mae_knn_sb2), convert_to_hms(mae_llm_sb2))
res_sb2['EA'] = (ea_usr_sb2, ea_heu_sb2, ea_dt_sb2, ea_rnp_sb2, ea_knn_sb2, ea_llm_sb2)
res_sb2['MAPE'] = (mape_usr_sb2/1000, mape_heu_sb2/1000, mape_dt_sb2/1000, mape_rnp_sb2/1000, mape_knn_sb2/1000, mape_llm_sb2/1000)

res_sb2

res_sb3 = pd.DataFrame()
res_sb3['Workload'] = ('H3', 'H3', 'H3', 'H3', 'H3', 'H3')
res_sb3['Method'] = ('User', 'Heuristic', 'DT', 'RNP', 'KNN', 'ORA')
res_sb3['MAE'] = (mae_usr_sb3, mae_heu_sb3, mae_dt_sb3, mae_rnp_sb3, mae_knn_sb3, mae_llm_sb3)
res_sb3['Average Error'] = (convert_to_hms(mae_usr_sb3), convert_to_hms(mae_heu_sb3), convert_to_hms(mae_dt_sb3), convert_to_hms(mae_rnp_sb3), convert_to_hms(mae_knn_sb3), convert_to_hms(mae_llm_sb3))
res_sb3['EA'] = (ea_usr_sb3, ea_heu_sb3, ea_dt_sb3, ea_rnp_sb3, ea_knn_sb3, ea_llm_sb3)
res_sb3['MAPE'] = (mape_usr_sb3/1000, mape_heu_sb3/1000, mape_dt_sb3/1000, mape_rnp_sb3/1000, mape_knn_sb3/1000, mape_llm_sb3/1000)

res_sb3

res_sb4 = pd.DataFrame()
res_sb4['Workload'] = ('H4', 'H4', 'H4', 'H4', 'H4', 'H4')
res_sb4['Method'] = ('User', 'Heuristic', 'DT', 'RNP', 'KNN', 'ORA')
res_sb4['MAE'] = (mae_usr_sb4, mae_heu_sb4, mae_dt_sb4, mae_rnp_sb4, mae_knn_sb4, mae_llm_sb4)
res_sb4['Average Error'] = (convert_to_hms(mae_usr_sb4), convert_to_hms(mae_heu_sb4), convert_to_hms(mae_dt_sb4), convert_to_hms(mae_rnp_sb4), convert_to_hms(mae_knn_sb4), convert_to_hms(mae_llm_sb4))
res_sb4['EA'] = (ea_usr_sb4, ea_heu_sb4, ea_dt_sb4, ea_rnp_sb4, ea_knn_sb4, ea_llm_sb4)
res_sb4['MAPE'] = (mape_usr_sb4/1000, mape_heu_sb4/1000, mape_dt_sb4/1000, mape_rnp_sb4/1000, mape_knn_sb4/1000, mape_llm_sb4/1000)

res_sb4

res_sb5 = pd.DataFrame()
res_sb5['Workload'] = ('H5', 'H5', 'H5', 'H5', 'H5', 'H5')
res_sb5['Method'] = ('User', 'Heuristic', 'DT', 'RNP', 'KNN', 'ORA')
res_sb5['MAE'] = (mae_usr_sb5, mae_heu_sb5, mae_dt_sb5, mae_rnp_sb5, mae_knn_sb5, mae_llm_sb5)
res_sb5['Average Error'] = (convert_to_hms(mae_usr_sb5), convert_to_hms(mae_heu_sb5), convert_to_hms(mae_dt_sb5), convert_to_hms(mae_rnp_sb5), convert_to_hms(mae_knn_sb5), convert_to_hms(mae_llm_sb5))
res_sb5['EA'] = (ea_usr_sb5, ea_heu_sb5, ea_dt_sb5, ea_rnp_sb5, ea_knn_sb5, ea_llm_sb5)
res_sb5['MAPE'] = (mape_usr_sb5/1000, mape_heu_sb5/1000, mape_dt_sb5/1000, mape_rnp_sb5/1000, mape_knn_sb5/1000, mape_llm_sb5/1000)

res_sb5

results = pd.concat([res, res_avg, res_sb1, res_sb2, res_sb3, res_sb4, res_sb5], axis=0)
results = results.reset_index(drop=True)
results

results.to_csv("metrics/pm100_9k_metrics.csv", index=False)
